Refactor cronjob options to replace WithEnableSeconds with WithCronSeconds
